Extract platform-specific code from ostd into PCI component
