+++ kernel/comps/pci/Cargo.toml
@@ -8,10 +8,14 @@ edition = "2021"
 [dependencies]
 ostd = { path = "../../../ostd" }
 component = { path = "../../libs/comp-sys/component" }
+align_ext = { path = "../../../ostd/libs/align_ext" }
 log = "0.4"
 spin = "0.9.4"
 cfg-if = "1.0"
 bitflags = "1.3"
+
+[target.loongarch64-unknown-none-softfloat.dependencies]
+fdt = { version = "0.1.5", features = ["pretty-printing"] }
 
 [lints]
 workspace = true
